+++ apps/bridge/components/NoCookies.tsx
@@ -1,13 +1,13 @@
 import { useEffect, useState } from "react";
+import { isInIframe } from '@layerswap/utils/common'
 import MessageComponent from "./MessageComponent";
-import inIframe from "./utils/inIframe";
 import Link from "next/link";
 
 function NoCookies(props) {
     const [embedded, setEmbedded] = useState<boolean>()
 
     useEffect(() => {
-        setEmbedded(inIframe())
+        setEmbedded(isInIframe())
     }, [])
 
     return (