+++ packages/utils/src/knownIds.ts
@@ -172,6 +172,10 @@ export default class KnownInternalNames {
 
         public static readonly TronTestnet: string = "TRON_TESTNET"
 
+        public static readonly StellarMainnet: string = "STELLAR_MAINNET"
+
+        public static readonly StellarTestnet: string = "STELLAR_TESTNET"
+
         public static readonly EclipseTestnet: string = "ECLIPSE_TESTNET"
 
         public static readonly EclipseMainnet: string = "ECLIPSE_MAINNET"