+++ packages/widget/core/src/components/utils/groupBy.ts
@@ -1,6 +1,1 @@
-export const groupBy = <T, K extends keyof any>(list: T[], getKey: (item: T) => K) => list.reduce((previous, currentItem) => {
-    const group = getKey(currentItem);
-    if (!previous[group]) previous[group] = [];
-    previous[group].push(currentItem);
-    return previous;
-}, {} as Record<K, T[]>);
+export { groupBy } from '@layerswap/utils/common'